requirements.txt
# --- Core Orchestration & Logic ---
langgraph>=0.4.0
langchain-openai>=1.1.0

# --- The Knowledge Layer (LlamaIndex) ---
# Better for Pinecone RAG than basic LangChain-Pinecone
llama-index>=0.12.0
llama-index-vector-stores-pinecone>=0.4.0
llama-index-embeddings-openai>=0.3.0

# --- Local Evaluation & Hugging Face ---
# Needed for your Phi-3 Grader Node
langchain-huggingface>=1.2.0
transformers>=4.48.0
torch>=2.5.0
bitsandbytes>=0.49.0
accelerate>=1.3.0

# --- Infrastructure & Data ---
pinecone-client>=5.4.0
psycopg2-binary>=2.9.10
redis>=5.2.0
celery>=5.4.0

# --- API & Validation ---
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
requests>=2.31.0  # For external API calls (FMP, SEC EDGAR)